requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
vk-api==11.9.9
python-telegram-bot==20.7
scikit-learn>=1.4.0
numpy>=1.26.0
nltk==3.8.1
dostoevsky==0.6.0
schedule==1.2.0
python-dotenv==1.0.0
lxml==5.0.0
cssselect==1.2.0

# RSS parsing
feedparser>=6.0.10

# Telegram User API
telethon>=1.34.0

# Proxy support
requests[socks]>=2.31.0
PySocks>=1.7.1
